application: generate node UUID on startup
Upon first starting up the storage layer, Redpanda will now check to see
if a node UUID exists in the controller keyspace of the key-value store.
If not, it will generate one and persist it to disk, and make it
available on all shards.

This will be used as a unique identifier when registering a node
(particularly important in the future when we auto-assign node IDs).